LPS-80122 contacts-uad - autogenerated
